MH-12877: Add new modal to edit multiple scheduled events at once
This PR adds a new modal dialog, accessible via the “Action” button in
the events overview. It only works for events that are scheduled, and
allows editing of such events using a simple wizard. It’s possible to
modify some of the metadata attributes as well as the scheduling
information.

Note that the “start date” is not editable, because it’s unclear what
that would do to multiple events. Also, you can only edit the start
time and the end time, not the duration. Again, this hopefully reduces
confusion.

This work was sponsored by SWITCH.

Co-authored-by: Kim Rinnewitz <krinnewitz@plapadoo.de>