MH-12879, Default location of paella configuration
The paella player lacks any default configuration for the location of
its main configuration files and is thus failing if this key is not
explicitly set, even if the default configuration file is provided.

This patch adds a default to `${karaf.etc}/paella` which is already
being used for the default configuration file.