MH-12889 Intuitive Merging of Video Segments
This patch changes the behavior when removing segments in the video
editor. Previously, the free space resulting from removing a segment was
taken up by extending the previous segment. Now, the subsequent segment
is extended instead, which leads to a more intuitive behavior. This is
especially the case when editing a video from left to right (from start
to end), which should be the most common use case.

For the last segment, the behavior is different: it has to be merged
with the previous segment since a subsequent one doesn't exist.