MH-12911: Override default cheat sheet and translate descriptions
For this, the description argument in the hotkey service
was removed. Instead of it, the hotkey service generates
a translation key from the key identifier and passes it
to `angular-hotkeys`.

I also used this opportunity to unify the capitalization
of all the descriptions and th fox some code formatting
inconsistencies.

The shortcuts in the cheat sheet are also grouped by
the first part of their key identifier.