MH-12926, Prevent cluttering of logs by invalid access
Every invalid access through domains not specifically mapped to an
organization will be logged in Opencast which makes it easy for an
external attacker to generate a huge amount of log files.

In production systems, we have seen that this caused issues with people
probing for default path's like /phpmyadmin, …

Since the warning is actually not indicating any issue with
Opencast–these requests should not be served–this patch downgrades the
log message to debug.