MH-12941, Gracefully handle empty flavors
While a media package element's flavor *should* always be set in the
scheduler, when it is not, the ingest for that event will fail. If the
capture agent which did record that event dutifully tries to upload the
event again later, this might force the ingest service repeatedly into
an error state.

This patch fixes the issue by not trying to merge elements with unset
flavors which makes sense in any case since the merge is based on
flavors and you cannot merge an element based on something which does
not exist.