MH-12966 Do not pre-select-from option in metadata property sheets
Open the Event Details modal causes the creation of a new snapshop under specific circumstances: In case that exactly one option is available, the pre-select-from directive will automatically select this option (to avoid the user to have to select the only available option manually).
In the case of the metadata property sheets, this behaviour seems neither correct nor appropriate: It is incorrect as for non-mandatory fields, there actually are two options: No option and the single available option. Inappropriate because most metadata fields can be configured and you would not configure a select for a single value, would you? I