MH-12974, Access denial to event for unprivileged user
Due to the incorrect usage of ``getAcl()``, unprivileged users may not
have access to events despite the ACL explicitly allowing them access.
Instead, `getActiveAcl()` should be used to evaluate the correct access
permissions for a given event.

Note that there are further, less critical issues with `getAcl()` as
mentioned on list [1] which will be dealt with separately with a patch
against `r/6.x`.

[1] https://groups.google.com/a/opencast.org/forum/#!topic/dev/CqGR1K3LNXE