MH-12970, Senseless XACML parsing
The method `getAclFromInputStream(…)` should parse an XACML file from a
given input stream and return the ACL object. The current implementation
internally parses the input stream, catching possible exceptions and
packing the result in an `Option` only to unpack it again right
afterwards which throws an exception in case the exception had been
caught before. Then the result is packed in a `Tuple` with a statically
set `AclScope` which, of course, is immediately unpacked again with the
scope just being discarded.

Finally, the one usage of this would break if there was any problem with
the ACL instead of handling this gracefully, falling back to a strict
non-permissive access rule.

This patch replaces all this mess with a straightforward implementation
of parsing the ACL from the given XACML input stream.