MH-12992: Trigger conflict check in “Edit scheduled” on “Next”
Conflict checking can be fairly slow on large systems with lots of
scheduled events. Currently, when you select a few of those events and
open the "Edit scheduled events" tab, changing a single item, such as
the start time, will already trigger a conflict check. If this takes
very long then, it's bad for usability.

This commit changes the behavior, so conflict checking is done when
you press the “Next” button in the wizard.

This work was sponsored by SWITCH.