MH-13003: Implement `Event#hasRecordingStarted` more reliably
After a scheduled event has been recorded, the index service
keeps the recording start date for the scheduler and the start date
in the meta data fields in sync, which makes renders unfit to
unambiguously decide whether an events **actual** recording start time
is in the past.

Here, we use the recording status instead, which is hopefully more reliable.