MH-13047, Document using Nginx for HTTPS
Our current documentation is not very helpful when it comes to
explaining the use of Nginx (or any other reverse proxy) with Opencast to
provide a proper HTTPS set-up despite this being the de-facto standard
for most Opencast installations.

This patch adds that documentation, while improving the general
structure of the HTTPS documentation and making it more important in the
general table of contents.

Note that the Nginx example included here is a minimal set-up, not
focused on security. A few general security hints will probably follow.