MH-13064, Encoding profile mimetypes are mostly ignored
Encoding profiles let you define mime types for the output of an
operation (e.g. video conversion) but these are mostly ignored by the
composer since an inspection is run on the output which will also set
the files mime type.

That behavior makes the mime type specifications in encoding profiles
not only a configuration overhead but also introduces unexpected
behavior since sometimes it is been used and sometimes it is not.

In fact, the only place a mime type needs to be specified is when
Opencast generates images since inspection is implemented to be run on
tracks instead of attachments. But for that, Opencast's MimeTypes class
has another way of automatically determining the mime type.

This patch completely removes mime type specifications from encoding
profiles. Note though, that specifying one will not cause an error since
encoding profiles support extensions. But it will have no effect.