MH-13067, Configuration panel does not work for default workflow
If a default workflow is configured in `custopm.properties` and hence
automatically selected in the “New Event” wizard, all user input from
the configuration panel is discarded and the default settings are used
instead.

This can be easily tested by configuring the default workflow:

    org.opencastproject.workflow.default.definition=schedule-and-upload

…and selecting the first two options in the configuration panel of the
“New Event” wizard. The summary will already show these values to be set
to false (despite of them being selected earlier).

This patch ensures the configuration panel is looked up when updating
the configuration to ensure the used configuration is always up-to-date.