MH-13079, Introduce REST Interface for AssetManager Properties
Asset manager properties allow for an event based key-value store under
a given namespace. This is currently used, for example, by the workflow
service which stores workflow variables passed to workflows as user
input to ensure they are available for a new workflow as well.

One problem of these properties is that they are never exposed, making
them extremely hard to access for things like a migration from
file-based workflow variable storage to asset manager based storage.

This patch aims to provide a basic API to read and write properties. It
is especially focused on workflow variables.