MH-13095, Add linter (checkstyle) for JavaScript
For a long time, Opencast has an automated style checker for all Java
code included in the project which ensured a mostly homogeneous style.
For JavaScript, we talked about introducing checks for a while but never
made any progress.

This patch introduces eslint as example linter for the runtime-info-ui
module. From here, this can be extended to other modules. Rules are not
configured to be very strict. The most notably enforcement is probably
the 2 space indention rule.