MH-13099, Warn when default credentials are being used
Using default credentials in production is bad and users should never do
that but we have still seen that happening.

This patch will cause Opencast to print a big warning if the passwords
are not changed. Everything will run as usual (important for testing)
but it's now more likely that users will notice they have a problem.