MH-13134, Concurrency issue when loading admin interface statistics
Opencast 6 makes the statistics counter in the admin interface
configurable. This can cause problems when the configuration is not yet
fully applied the first time, the counters are loaded as the refresh
interval is then never set and the counter values fail to load.

This patch simply adds the usual refresh if the configuration is not yet
properly set, ensuring that the values are loaded eventually.