Bump angular-ui-sortable, adapting build pipeline
This will update angular-ui-sortable to the highest version that is
still compatible with the jQuery version in use. By doing this via
bower, any direct dependencies will be added via bower as well to
make use of the build pipeline.

As a side-effect, manually injected scripts in the head tag having
dependencies that are moved to the end of the body tag will be
added to the build pipeline as well, allowing for the load order
to stay consistent.

Resolves: MH-13054