Fix initialisation of stats and improve its destruction
The stats directive attempted to fetch the stats data which cannot be done at this point of time as the stats service has not yet been configured. As we cannot fetch data while not data source has been configured, let's start polling after the stats service has been configured.
The destructor has been moved to the directive to have this logic with the stats logic rather than the events controller.