MH-13142, Error parsing non-existent schedule
In some situations (e.g. opening the event details), the admin interface
tries to load and parse scheduling information, even if none exist,
causing the JSON parser to throw an exception since the endpoint does
not return JSON.

This patch stops the data handling if the response code indicates that
there are no information.