MH-13149, Timed tiered storage test fails on fast systems
The asset manager tests are testing that certain snapshots are created
within a certain timespan. This may fail on faster systems since it is
not guaranteed that the measured times are not equal.

For example, if the first snapshot gets the same time as the measurement
right before it, the request for snapshots until that date returns a
snapshot while none is expected. That causes the test to fail.

This regularly happened on one of my development set-ups.

This patch ensures the times are not equal by enforcing a millisecond
(the granularity of Date) sleep between the measured operations.