MH-13169 Update bibliographic metadata when technical metadata changes
When scheduling a new event, the "bibliographic" meta data (as it can
be seen under event details -> meta data) is initialized with the
technical meta data specified while scheduling the event. However, when
changing the technical meta data of an exsiting event, the bibliographic
meta data is not updated accordingly. This is quite confusing for the
user. What makes things worse is the fact that the events table reflects
the change to the start/end date while it doesn't display the new
location.

To overcome this confusing behavior, this patch updates the
bibliographic meta data whenever the technical meta data is changed.
Changing the bibliographic meta data, however, doesn't affect the
technical meta data.