MH-13175, Remove Apache Tika for Generating Mimetypes
Opencast comes with a lot of methods to determine mime types for
content. While the number of methods has been reduced over the years,
there are still multiple ways and Apache Tika is one of them.

By removing Tika and replacing it with the MimeTypes class, we not only
reduce dependencies but also reduce OSGI services we run.