MH-13162 Show all series in edit-scheduled-events
Show all series in the edit-scheduled-events modal instead of only the
first 100. Use the listprovider endpoint since the request is
significantly faster compared to the series endpoint of the admin UI.
Only switch to the second tab once the series have been loaded,
otherwise the dropdown will be empty. Update the rest
docs of the series endpoint.