MH-13191 Improve performance of event deletion
When deleting events from the scheduler service, some unnecessary
queries related to properties are executed:

- Since we know, that we are deleting the whole episode, we do not have
to check whether there are snapshots for the properties to be deleted. I
am not sure anyway, why this is necessary at all.

- Since we know, that we take care of deleting all properties related to
the episode, we do not have to remove orphaned properties (because we are
not producing any).

It turned out to be quiet hard to control the behavior of the code
executed in AbstractADeleteQuery.runQueries() and I am not happy with
the flag attached to DeleteQueryContribution. However, I could not find a
better solution in the available time.

All in all, these changes seem to improve event deletion by about 50 or
60%, according to our measurements.