MH-13186, Introduce ACL Merge Modes
Access control lists can be specified both on series and on episode
level. This means that multiple rule sets can be attaches to an episode
which is part of a series. With this patch, Opencast can now handle this
in multiple ways. If multiple lists are attached, the following modes
define Opencast's behavior:

Merge Mode “override” (Default)
-------------------------------

The episode ACL takes precedence over the series ACL. This means that
the series ACL will be completely ignored as soon as the episode has an
ACL, no matter what rules are set in either. This allows users to define
general rules for a series which can be completely redefined on an
episode and which are not influenced by changes later made to the
series.  This is also a very simple rule and thus easy to understand.

Example:

|         |ROLE1 |       |ROLE2 |       |ROLE3 |       |
|---------|------|-------|------|-------|------|-------|
|         |*read*|*write*|*read*|*write*|*read*|*write*|
|*series* |allow |allow  |allow |allow  |      |       |
|*episode*|      |       |allow |       |allow |       |
|*active* |      |       |allow |       |allow |       |

Merge Mode “roles”
------------------

Series and episode ACLs are merged based on the roles defined within. If
both the series and the episode define a rule for a specific role (user
or group), the episode's rule takes precedence. Rules for roles defined
in one ACL only are always part of the resulting active ACL.

Example:

|         |ROLE1 |       |ROLE2 |       |ROLE3 |       |
|---------|------|-------|------|-------|------|-------|
|         |*read*|*write*|*read*|*write*|*read*|*write*|
|*series* |allow |allow  |allow |allow  |      |       |
|*episode*|      |       |allow |       |allow |       |
|*active* |allow |allow  |allow |       |allow |       |

Merge Mode “actions”
--------------------

ACLs are merged based on the actions (read, write, …) contained within
both ACLs. If a rule is specified for a tuple of role and action in both
ACLs, the rule specified in the episode ACL takes precedence.

Example:

|         |ROLE1 |       |ROLE2 |       |ROLE3 |       |
|---------|------|-------|------|-------|------|-------|
|         |*read*|*write*|*read*|*write*|*read*|*write*|
|*series* |allow |allow  |allow |allow  |      |       |
|*episode*|      |       |allow |       |allow |       |
|*active* |allow |allow  |allow |allow  |allow |       |