MH-13185, Don't include test web server
For some reason, Opencast's common module included a whole web server as
well as some other components which were exclusively used for unit tests
but which were included in the distribution nonetheless.

This patch puts this common test code in its own internal library so it
can easily be included for testing purposes if necessary.