MH-13191 Improve performance of event deletion
The AssetManager used to "manually" control the relation between
snapshots and assets. I.e. assets actually have references to snapshots,
but this relationship was only handled on application level. On database
level, there was not explicitly declared relationship between assets and
snapshots.

So when deleting snapshots, the AssetManager had to first figure out,
which snapshots were going to be deleted by the given query and then
"manually" remove all assets referencing the snapshots to be deleted,
before actually deleting the snapshots.

With this patch, the relationship between assets and snapshots is
modeled explicitly on the database level by changing the shema
accordingly. This allows for ON DELETE CASCADE to be used resulting in
only one query being necessary when removing snapshots.