MH-13197, Separate Admin Interface Index
The search index of the admin interface is bundled in the same package
which also contains the REST endpoints and the frontend, resulting in a
re-start of a lot of components if you just want to deploy a simple
front-end fix.

This patch separated the index so that a replacement of the admin
interface bundle does not cause the index to re-start as well making
hot-deployments easier.

Note that commit a52a455 fixes a few problems with updating bundles
which will come handy if this is to be used.