MH-13192 Improve performance of list requests
The following requests are quite slow when asking for a large number of objects:

- GET /admin-ng/events
- GET /api/events
- GET /admin-ng/series
- GET /api/series

An analysis revealed that most of the time is spent creating JAXB
Unmarshaller objects for each single item. This can be sped up by
reusing the same Unmarshaller between events/series.

Note that this is not thread-safe, of course, but we’re inside a
single thread while deserializing, so this should be fine.