MH-13213, Separate External API Index
The search index of the external API is bundled in the same package
which also contains the REST endpoints, resulting in a re-start of a lot
of components if you just want to deploy a simple fix.

This patch separated the index so that a replacement of the external-api
bundle does not cause the index to re-start making hot-deployments
easier.