MH-13216, Fix Documentation Pages
Looking at the output of a documentation build using a recent version
mkdocs reveals a number of issues in the documentation:

- Use of now deprecated mkdocs configuration
- Several invalid linked files
- Several files not properly included

This patch fixes those problems and icludes a general test for those
problems as part of the CI on Travis.

Note that this may make the documentation projects incompatible to old
versions of mkdocs. Use `pip` to get a recent version if necessary.