MH-13224, Drop commons-beanutils
Opencast's REST docs support macro expansion in documentation strings
via commons-beanutils (basically a wrapper around Java's Reflection and
Introspection APIs). This old library (new really new version is
available for this) causes a few conflicts with the Karaf 4.2 update.

Since this feature is almost exclusively used for setting string
constants which is supported directly by simply using Java constants
instead (and with a lot less black magic involved), I've went for
dropping the feature to solve the conflicts.

This patch drops the macro expansion, completely removes the usage of
commons-beanutils and sets (and in some cases fixes) a number of default
values for REST endpoints.