MH-13228, Homogeneous Width of Shortcut Icons
The icons indicating which key to press in the shortcut overview of the
admin interface all have a different width based on the letters used
which looks add.

This patch sets a minimum with which should create a more homogeneous
design by default while not restricting the width for special cases.