MH-13284, Update Elasticsearch to 2.x
As a first step in the process of bringing Elasticsearch to the latest
stable version, this patch updates Elasticsearch and its client
libraries to the latest 2.x version.

Unnecessary changes in how Opencast and Elasticsearch are avoided if
possible to make the update as smooth as possible though some queries
had to be replaces and the structure had to be adjusted slightly.

*This requires an index rebuild*