MH-13285, Display workflow description
…when selecting a workflow.

Since a description could be a very long text and will likely be ignored
by an admin who is familiar with the workflow, the description it is
contained in a div that becomes "expandable"/"collapsible" when it
overflows the defined max-height.

An empty line in the input will be formatted to a line break, while a
single line break will be ignored. (Single line breaks are keeping the
xml clean and compact and shouldn't be displayed).

Signed-off-by: Lars Kiesow <lkiesow@uos.de>

- added description
- squashed nonsense commits