MH-13310: Simplify `AQueryBuilderImpl#always`
The previous implementation basically created a predicate of the form
`q == q`, where `q` is some query variable (here `Q_SNAPSHOT`),
to create a predicate that is always true.
This works, but introduces `q` as a variable in the `WHERE` clasue
of every query that uses this combinator, which in turn requires
a matching declaration in the `FROM` clause, even though any query
in particular might not even need `q`.

The new implementation is still a bit roundabout. I would have
preferred the empty condition to just look like `true` when rendered
to SQL, but that is not valid JPQL. `true = true` works, though.