MH-13310: Expose deletion callback in the `ADeleteQuery` interface
Instead of just returning how many elements were deleted,
the delete queries from the asset manager can now call you back
about which specific snapshots and episodes were deleted.
This mechanism existed before, hidden in `AbstractADeleteQuery`
as an implementation detail, but it is a useful mechanism
that might as well be exposed in the interface for everyone
to use.