MH-13322, Avoid . in Elasticsearch Field Names
Elasticsearch ≥ 2 does not allow '.' to be used in any field names.
Right now, this can happen with the stored capture agent properties in
the events index. Luckily we do not seem to use those data at all right
now.

This patch removes setting those data and introduces an additional check
for those fields.