MH-13355 Increase the default timeout for TrustedHttpClientImpl from 1min to 5min
Inter-server operations like ingest -> admin to start a new workflow on an ingested mediapackage use a POST to start a workflow.

Under some conditions, this can take longer than 60s, in which case the calling node (ingest server) thinks the request has failed, whereas it actually executes fine.

This increases the default timeout for HttpTrustedClient to 5 minutes so it doesn't fail prematurely, which will make Opencast a little more robust and avoid services being flagged as WARN or ERROR unnecessarily.