Revert #523: Special handling of asset manager event removal
This patch reverts pull request #523 which introduced a special query
option to indicate that a query will remove a whole media package which
would disable a set of checks and cleanup operations.

Since pull request #696 (alone with the other asset manager performance
fixes) removed the whole expensive handling which was circumvented, this
now only removes some left-over complexity which had no gain any longer.

Note that the performance patches are not yet merged into `develop`
which is intentional since they cause merge conflicts with pull
request #523 meaning that the merge should be way cleaner after this is
merged.