MH-13346 Thumbnail generation from Admin UI can fail on longer videos
Extracting an image requires the source track to be download into the workspace. This is a very slow operation for large files that can cause problems in case of recordings with unusual long duration. This commit introduces the configuration property thumbnail.preview.mode which defaults to 'partial'. In that mode, the thumbnail preview will not be automatically extracted by the video editor, but all the logic for thumbnails is still active.