MH-13439, Dynamic Player Redirect
This patch adds a `/watch/<id>` endpoint as a universal place for
Opencast players, regardless of which specific player is configured to
be used right now. The endpoint will redirect users to the correct place
using an HTTP redirect.

Furthermore, this patch integrates the new endpoint into all relevant
Opencast components like the publication channels and the engage tools.

Note that this behavior is similar to the JavaScript redirect included
in `/engage/ui/watch.html` in commit aac56d7 which was later partially
broken by commit 23dbfa9 causing the direct player links to be written
into the media packages publication elements directly.

But unlike the old mechanism, the new endpoint does not require the web
browser to render the page before issuing the redirect or even load any
additional resources.

To sum up what this patch does:

- Add a new player redirection endpoint

- Use this endpoint wherever a player link is required to unify the
  behavior

- Update the player settings in the organization configuration file

- Deprecate the old `/engage/ui/watch.html`

- Provide instructions to migrate all old links if desired